converted some ints to type long because during octal -> decimal conversion the number exceeded the max int size and was producing very incorrect results. Once the number has been converted back to a decimal its type is restored to int due to the return type of the function octal_to_decimal

+++ my_libc_funcs/my_atoi.c
@@ -1,7 +1,7 @@
 #include "../my_tar.h"
 
-int		my_atoi(char *num) {
-	unsigned int result = 0;
+long	my_atoi(char *num) {
+	long result = 0;
 	int sign = 1;
 	int i = 0;
